requirements.txt
groq>=0.9.0
google-generativeai>=0.8.0
flask>=3.0.0
flask-cors>=4.0.0
